Refactor designer page and add hooks
